+++ src/index.ts
@@ -1,20 +1,8 @@
-import * as admin from 'firebase-admin';
+import { FirebaseHelper } from './firebase';
 import { FirestoreHelper } from './firestore';
 
+const firebaseHelper = new FirebaseHelper();
 const firestoreHelper = new FirestoreHelper();
 
-/**
- * Initialize Firebase App
- * 
- * @param {any} serviceAccount 
- * @param {any} databaseURL
- */
-exports.initializeApp = function (serviceAccount: string, databaseURL: string) {
-    admin.initializeApp({
-        credential: admin.credential.cert(serviceAccount),
-        databaseURL: databaseURL
-    })
-    return {'firestore': admin.firestore()};    
-}
-
+exports.firebase = firebaseHelper;
 exports.firestore = firestoreHelper;